Catch errors that are thrown in hooks
Prior to this commit, if you have an ^:after-load fn that can throw an error (E.g
reagent failing to render because you're rendering something like
`[[]]`) figwheel will fail to reload after fixing the error. You have
to click refresh in the browser for figwheel to fix itself or figure out
that you need to swallow errors in your ^:after-load fn.